+++ AlbaGo/src/main/java/com/albago/service/NoticeService.java
@@ -6,5 +6,5 @@ import com.albago.domain.NoticeVO;
 
 public interface NoticeService {
 
-	public List<NoticeVO> getList(int c_code);
+	public List<NoticeVO> getList(long c_code);
 }
